# make polars plugin use mimalloc