--- a/crates/nu_plugin_polars/src/main.rs
+++ b/crates/nu_plugin_polars/src/main.rs
@@ -1,6 +1,9 @@
 use nu_plugin::{serve_plugin, MsgPackSerializer};
 use nu_plugin_polars::PolarsPlugin;
 
+#[global_allocator]
+static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
+
 fn main() {
     serve_plugin(&PolarsPlugin::default(), MsgPackSerializer {})
 }