@maxim-uvarov 2024-05-25
mimalloc won by 10ms on average

Using jemalloc
```
> bench -n 100 {polars open Data7602DescendingYearOrder.csv | polars group-by year | polars agg (polars col geo_count | polars sum) | polars collect | null}
╭───────┬───────────────────────╮
│ mean  │ 112ms 250µs 864ns     │
│ min   │ 105ms 303µs 334ns     │
│ max   │ 126ms 30µs            │
│ std   │ 3ms 983µs 723ns       │
```

Using mimalloc
```
> bench -n 100 {polars open Data7602DescendingYearOrder.csv | polars group-by year | polars agg (polars col geo_count | polars sum) | polars collect | null} | table
╭───────┬───────────────────╮
│ mean  │ 101ms 195µs 696ns │
│ min   │ 94ms 697µs 666ns  │
│ max   │ 117ms 242µs 833ns │
│ std   │ 4ms 178µs 744ns   │
```